Struct epub::doc::EpubDoc

pub struct EpubDoc {
    pub spine: Vec<String>,
    pub resources: HashMap<String, (String, String)>,
    pub root_base: String,
    pub root_file: String,
    // some fields omitted
}

Struct to control the epub document

Fields
spine: Vec<String>

epub spine ids

resources: HashMap<String, (String, String)>

resource id -> name

root_base: String

root file base path

root_file: String

root file full path

Methods

impl EpubDoc

fn new(path: &str) -> Result<EpubDoc, Box<Error>>

Opens the epub file in path.

Initialize some internal variables to be able to access to the epub spine definition and to navigate trhough the epub.

Errors

Returns an error if the epub is broken or if the file doesn't exists.